+++ Projects/TDPresentation/Sources/AppFlow/MainFlow/Diary/FocusCalendar/FocusCalendarViewController.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+final class FocusCalendarViewController: BaseViewController<BaseView> {
+    
+}
