+++ Projects/TDPresentation/Sources/AppFlow/MainFlow/Diary/FocusCalendar/FocusCalendarViewModel.swift
@@ -1,0 +1,21 @@
+import Combine
+import Foundation
+
+final class FocusCalendarViewModel: BaseViewModel {
+    enum Input {
+        
+    }
+    
+    enum Output {
+        
+    }
+    
+    private let output = PassthroughSubject<Output, Never>()
+    private var cancellables = Set<AnyCancellable>()
+    
+    func transform(input: AnyPublisher<Input, Never>) -> AnyPublisher<Output, Never> {
+        
+        return output.eraseToAnyPublisher()
+    }
+}
+
